Extract common logic to use-database-connection hook
